GP-4881: Prevent PE Data Directory parse failures from making the import fail
