+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/bin/format/mz/DOSHeader.java
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -150,8 +150,8 @@ public class DOSHeader extends OldDOSHeader {
 		if (e_lfanew >= 0 && e_lfanew <= 0x1000000) {
 			try {
 				NTHeader ntHeader =
-					new NTHeader(reader, e_lfanew, SectionLayout.FILE, false, false);
-				if (ntHeader != null && ntHeader.getOptionalHeader() != null) {
+					new NTHeader(reader, e_lfanew, SectionLayout.FILE, false);
+				if (ntHeader.getOptionalHeader() != null) {
 					return true;
 				}
 			}
